% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf.R
\name{is_tf_available}
\alias{is_tf_available}
\title{Test to see if tensorflow is available}
\usage{
is_tf_available()
}
\value{
A logical
}
\description{
Test to see if tensorflow is available
}
\examples{
is_tf_available()
}
