% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.recipe}
\alias{tidy.recipe}
\title{Tidy the Result of a Recipe}
\description{
\code{tidy} will return a data frame that contains information regarding a recipe
or operation within the recipe (when a \code{tidy} method for the operation
exists). See \link[recipes:tidy.recipe]{recipes::tidy.recipe} for more information.
}
