% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunable.R
\name{tunable.step_embed}
\alias{tunable.step_embed}
\alias{tunable.step_umap}
\alias{tunable.step_woe}
\alias{tunable.step_discretize_xgb}
\alias{tunable.step_discretize_cart}
\title{tunable methods for embed}
\usage{
tunable.step_embed(x, ...)

tunable.step_umap(x, ...)

tunable.step_woe(x, ...)

tunable.step_discretize_xgb(x, ...)

tunable.step_discretize_cart(x, ...)
}
\arguments{
\item{x}{A recipe step object}

\item{...}{Not used.}
}
\value{
A tibble object.
}
\description{
These functions define what parameters \emph{can} be tuned for specific steps.
They also define the recommended objects from the \code{dials} package that can
be used to generate new parameter values and other characteristics.
}
\keyword{internal}
