% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed-package.R
\docType{package}
\name{embed-package}
\alias{embed}
\alias{embed-package}
\title{embed: Extra Recipes for Encoding Predictors}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Predictors can be converted to one or more numeric representations using a variety of methods. Effect encodings using simple generalized linear models <arXiv:1611.09477> or nonlinear models <arXiv:1604.06737> can be used. There are also functions for dimension reduction and other approaches.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://embed.tidymodels.org}
  \item \url{https://github.com/tidymodels/embed}
  \item Report bugs at \url{https://github.com/tidymodels/embed/issues}
}

}
\author{
\strong{Maintainer}: Max Kuhn \email{max@rstudio.com} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
