% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R, R/discretize_cart.R,
%   R/discretize_xgb.R, R/glm.R, R/hash.R, R/lme.R, R/tf.R, R/umap.R, R/woe.R
\name{required_pkgs.step_lencode_bayes}
\alias{required_pkgs.step_lencode_bayes}
\alias{required_pkgs.step_discretize_cart}
\alias{required_pkgs.step_discretize_xgb}
\alias{required_pkgs.step_lencode_glm}
\alias{required_pkgs.step_feature_hash}
\alias{required_pkgs.step_lencode_mixed}
\alias{required_pkgs.step_embed}
\alias{required_pkgs.step_umap}
\alias{required_pkgs.step_woe}
\title{S3 methods for tracking which additional packages are needed for steps.}
\usage{
\method{required_pkgs}{step_lencode_bayes}(x, ...)

\method{required_pkgs}{step_discretize_cart}(x, ...)

\method{required_pkgs}{step_discretize_xgb}(x, ...)

\method{required_pkgs}{step_lencode_glm}(x, ...)

\method{required_pkgs}{step_feature_hash}(x, ...)

\method{required_pkgs}{step_lencode_mixed}(x, ...)

\method{required_pkgs}{step_embed}(x, ...)

\method{required_pkgs}{step_umap}(x, ...)

\method{required_pkgs}{step_woe}(x, ...)
}
\arguments{
\item{x}{A recipe step}
}
\value{
A character vector
}
\description{
Recipe-adjacent packages always list themselves as a required package so that
the steps can function properly within parallel processing schemes.
}
\keyword{internal}
