% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_cart.R, R/collapse_stringdist.R,
%   R/discretize_cart.R, R/discretize_xgb.R, R/embed.R, R/lencode.R,
%   R/lencode_bayes.R, R/lencode_glm.R, R/lencode_mixed.R, R/pca_sparse.R,
%   R/pca_sparse_bayes.R, R/pca_truncated.R, R/umap.R, R/woe.R
\name{required_pkgs.step_collapse_cart}
\alias{required_pkgs.step_collapse_cart}
\alias{required_pkgs.step_collapse_stringdist}
\alias{required_pkgs.step_discretize_cart}
\alias{required_pkgs.step_discretize_xgb}
\alias{required_pkgs.step_embed}
\alias{required_pkgs.step_lencode}
\alias{required_pkgs.step_lencode_bayes}
\alias{required_pkgs.step_lencode_glm}
\alias{required_pkgs.step_lencode_mixed}
\alias{required_pkgs.step_pca_sparse}
\alias{required_pkgs.step_pca_sparse_bayes}
\alias{required_pkgs.step_pca_truncated}
\alias{required_pkgs.step_umap}
\alias{required_pkgs.step_woe}
\title{S3 methods for tracking which additional packages are needed for steps.}
\usage{
\method{required_pkgs}{step_collapse_cart}(x, ...)

\method{required_pkgs}{step_collapse_stringdist}(x, ...)

\method{required_pkgs}{step_discretize_cart}(x, ...)

\method{required_pkgs}{step_discretize_xgb}(x, ...)

\method{required_pkgs}{step_embed}(x, ...)

\method{required_pkgs}{step_lencode}(x, ...)

\method{required_pkgs}{step_lencode_bayes}(x, ...)

\method{required_pkgs}{step_lencode_glm}(x, ...)

\method{required_pkgs}{step_lencode_mixed}(x, ...)

\method{required_pkgs}{step_pca_sparse}(x, ...)

\method{required_pkgs}{step_pca_sparse_bayes}(x, ...)

\method{required_pkgs}{step_pca_truncated}(x, ...)

\method{required_pkgs}{step_umap}(x, ...)

\method{required_pkgs}{step_woe}(x, ...)
}
\arguments{
\item{x}{A recipe step}
}
\value{
A character vector
}
\description{
Recipe-adjacent packages always list themselves as a required package so that
the steps can function properly within parallel processing schemes.
}
\keyword{internal}
