% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_linear.R
\name{step_adjust_linear}
\alias{step_adjust_linear}
\alias{tidy.step_adjust_linear}
\title{Adjust variables using a linear model}
\usage{
step_adjust_linear(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  remove_vars = NULL,
  keep_vars = NULL,
  models = NULL,
  drop = c("remove", "both", "none"),
  skip = FALSE,
  id = rand_id("adjust_linear")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of
operations for this recipe.}

\item{...}{One or more selector functions to choose variables for this step.
See \code{\link[recipes:selections]{selections()}} for more details.}

\item{role}{Not used by this step since no new variables are created.}

\item{trained}{A logical to indicate if the quantities for preprocessing have
been estimated.}

\item{remove_vars}{One or more selector functions to choose variables to
residualize out. The predicted term-values for these variables are
subtracted from the outcome (\code{...}).}

\item{keep_vars}{One or more selector functions to choose variables to
\emph{not} residualize out.}

\item{models}{The \code{\link[stats:lm]{stats::lm()}} object is stored here once this preprocessing
step has be trained by \code{\link[recipes:prep]{recipes::prep()}}.}

\item{drop}{When \code{\link[recipes:bake]{recipes::bake()}} is called, should the \code{remove_vars}
variables be removed (\code{"remove"}; default), also the \code{keep_vars} variables
(\code{"both"}) or should no variables be removed (all kept; \code{"none"}).}

\item{skip}{A logical. Should the step be skipped when the recipe is baked by
\code{\link[recipes:bake]{bake()}}? While all operations are baked when \code{\link[recipes:prep]{prep()}} is run, some
operations may not be able to be conducted on new data (e.g. processing the
outcome variable(s)). Care should be taken when using \code{skip = TRUE} as it
may affect the computations for subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}
}
\value{
An updated version of \code{recipe} with the new step added to the
sequence of any existing operations.
}
\description{
\code{step_adjust_linear()} creates a \emph{specification} of a recipe step that will
adjust a variable or group of variables by linearly "residualizing out" other
variable(s).
}
\details{
For each selected variable, \code{step_adjust_linear()} fit a \emph{linear} model:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{lm(variable ~ remove_vars + keep_vars)
}\if{html}{\out{</div>}}

And then adjusts \code{variable} but subtracting from the it sum of the predicted
term-wise values of \code{remove_vars} (using
\code{\link[stats:predict.lm]{stats::predict.lm(term = remove_vars)}}. This is similar
to the functionality provided by \code{\link[limma:removeBatchEffect]{limma::removeBatchEffect()}}, and is
particularly useful for "removing" unwanted batch effects from
log-gen-expression outcomes associated with technical variables (possibly
without removing experimental design or grouping variables), but can be
applied to any situation where variables need to be adjusted as part of
pre-processing.

(Prior to model fitting, numerical \code{remove_vars} / \code{keep_vars} are centered
and factors are effect-encoded using \code{\link[stats:contr.sum]{stats::contr.sum()}}.)

Note that the original data will be replaced with the adjusted data, possibly
dropping the \code{remove_vars} / \code{keep_vars} (depending on the value of the \code{drop} argument).
}
\section{Tidying}{
When you \code{\link[recipes:tidy.recipe]{tidy()}} this step, a tibble is returned with
columns \code{variables}, \code{term}, \code{type}, \code{value}, and \code{id}:

\describe{
\item{variables}{character, the selectors or variables selected}
\item{term}{character, the variables to remove or keep / coefficient label}
\item{type}{character, either "remove" or "keep"}
\item{value}{numeric, the coefficient value for the term}
\item{id}{character, id of this step}
}
}

\section{Case weights}{


This step performs an supervised operation that can utilize case weights.
To use them, see the documentation in \link[recipes:case_weights]{recipes::case_weights} and the examples on
\code{tidymodels.org}.
}

\examples{
\dontshow{if (rlang::is_installed(c("modeldata", "ggplot2 (>= 4.0.0)"))) withAutoprint(\{ # examplesIf}

library(ggplot2)

data("penguins", package = "modeldata")
penguins <- na.omit(penguins)

p <- ggplot(penguins, aes(flipper_length_mm, body_mass_g, color = sex)) +
  geom_point(aes(shape = species)) +
  stat_ellipse() +
  labs(title = "No adjustment")

p


recipe <- recipe(body_mass_g ~ ., data = penguins) |>
  step_adjust_linear(
    flipper_length_mm,
    body_mass_g,
    remove_vars = vars(species),
    keep_vars = vars(sex),
    drop = "none" # keep all variables in the baked data
  )

p +
  (prep(recipe) |>
    bake(new_data = penguins)) +
  labs(title = "Adjustment for species")

\dontshow{\}) # examplesIf}
}
