% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize_cart.R, R/discretize_xgb.R,
%   R/embed.R, R/pca_sparse.R, R/pca_sparse_bayes.R, R/tunable.R, R/umap.R,
%   R/woe.R
\name{tunable.step_discretize_cart}
\alias{tunable.step_discretize_cart}
\alias{tunable.step_discretize_xgb}
\alias{tunable.step_embed}
\alias{tunable.step_pca_sparse}
\alias{tunable.step_pca_sparse_bayes}
\alias{tunable_embed}
\alias{tunable.step_umap}
\alias{tunable.step_woe}
\title{tunable methods for embed}
\usage{
\method{tunable}{step_discretize_cart}(x, ...)

\method{tunable}{step_discretize_xgb}(x, ...)

\method{tunable}{step_embed}(x, ...)

\method{tunable}{step_pca_sparse}(x, ...)

\method{tunable}{step_pca_sparse_bayes}(x, ...)

\method{tunable}{step_umap}(x, ...)

\method{tunable}{step_woe}(x, ...)
}
\arguments{
\item{x}{A recipe step object}

\item{...}{Not used.}
}
\value{
A tibble object.
}
\description{
These functions define what parameters \emph{can} be tuned for specific steps.
They also define the recommended objects from the \code{dials} package that can be
used to generate new parameter values and other characteristics.
}
\keyword{internal}
